% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{adapt_distribution}
\alias{adapt_distribution}
\title{Simmer generator distribution adaptor function}
\usage{
adapt_distribution(distribution_function)
}
\arguments{
\item{distribution_function}{distribution functions}
}
\value{
adapted generator functions, which has empty formal arguments and produces an element
}
\description{
Simmer generator distribution adaptor function
}
\keyword{internal}
