% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_policy.r
\name{adapt_policy_selection_function}
\alias{adapt_policy_selection_function}
\title{Wrapper for custom policy selection function}
\usage{
adapt_policy_selection_function(select_function, policy)
}
\description{
Wrapper for custom policy selection function
}
\keyword{internal}
