% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iti_entity.R
\name{add_iti_entities}
\alias{add_iti_entities}
\title{Add IT infrastructure entities based on provided list of config objects}
\usage{
add_iti_entities(.env, ..., .configs)
}
\arguments{
\item{.env}{IT infrastructure environment}

\item{...}{configs to be wrapped into typed list}

\item{.configs}{typed list of configs. This is primary list of configs to be used in case if dots and configs are both not empty}
}
\value{
enriched IT infrastructure environment
}
\description{
Add IT infrastructure entities based on provided list of config objects
}
