% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iti_entity.R
\name{add_iti_resource_entity}
\alias{add_iti_resource_entity}
\title{Wrapper around simmer::add_resource based on provided list of arguments}
\usage{
add_iti_resource_entity(.env, .config)
}
\arguments{
\item{.env}{IT infrastructure environment}

\item{.config}{config object or basically a list of arguments}
}
\value{
Enriched IT infrastructure environment
}
\description{
Wrapper around simmer::add_resource based on provided list of arguments
}
