% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.r
\name{add_scheduler}
\alias{add_scheduler}
\title{Add scheduler/load balancer}
\usage{
add_scheduler(.env, ..., .config)
}
\arguments{
\item{.env}{Simulation environment with name "IT infrastructure"}

\item{...}{Add resource parameters from simmer}

\item{.config}{scheduler config object which will be used as primary source of parameters for the resource}
}
\value{
Simulation environment with resource named "Scheduler"
}
\description{
Add scheduler/load balancer
}
