% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.r
\name{add_servers}
\alias{add_servers}
\title{Add a list of servers to IT infrastructure}
\usage{
add_servers(.env, .servers)
}
\arguments{
\item{.env}{the simulation environment as IT infrastructure}

\item{.servers}{server configurations}
}
\value{
environment with added server resources
}
\description{
Define resources with predefined name "Server"
}
