% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{as.data.frame.density}
\alias{as.data.frame.density}
\title{S3 method for as.data.frame when 'density' type is provided}
\usage{
\method{as.data.frame}{density}(.density)
}
\arguments{
\item{.density}{Density object from stats package}
}
\value{
data frame which contains only 'x' and 'y' values from 'density' object
}
\description{
S3 method for as.data.frame when 'density' type is provided
}
