% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulation_result.r
\name{as.policy_simulation_result}
\alias{as.policy_simulation_result}
\title{Convert object to policy_simulation_result}
\usage{
as.policy_simulation_result(x)
}
\arguments{
\item{x}{Object to convert}
}
\value{
New 'policy_simulation_result' object
}
\description{
Convert object to policy_simulation_result
}
