% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{change_scheduler_handle_time}
\alias{change_scheduler_handle_time}
\title{Change timeout handle time for resource}
\usage{
change_scheduler_handle_time(.trj, ...)
}
\arguments{
\item{.trj}{request trajectory}

\item{...}{
  Arguments passed on to \code{\link[=change_timeout_for]{change_timeout_for}}
  \describe{
    \item{\code{resource}}{resource name to lookup}
    \item{\code{handle_function}}{new timeout handle function}
  }}
}
\description{
Change timeout wrapper for scheduler
}
