% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{change_timeout_for}
\alias{change_timeout_for}
\title{Change timeout handle time for resource}
\usage{
change_timeout_for(.trj, resource, handle_function)
}
\arguments{
\item{.trj}{request trajectory}

\item{resource}{resource name to lookup}

\item{handle_function}{new timeout handle function}
}
\value{
updated trajectory
}
\description{
Change timeout handle time for resource
}
