% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{config.repeat}
\alias{config.repeat}
\title{Create list of config objects with the same parameters and different names}
\usage{
config.repeat(n = 1, config_type = c("server", "scheduler"), ...)
}
\arguments{
\item{n}{Number of items}

\item{config_type}{Type of config to generate}

\item{...}{Parameters of the config}
}
\value{
list of config objects
}
\description{
Create list of config objects with the same parameters and different names
}
