% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{config.repeat.scheduler}
\alias{config.repeat.scheduler}
\title{Repeat function handler for scheduler_config}
\usage{
config.repeat.scheduler(n = 1, ...)
}
\arguments{
\item{n}{Number of items}

\item{...}{Additional arguments}
}
\value{
list of scheduler_config object with the same parameters
}
\description{
Repeat function handler for scheduler_config
}
