% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{config.repeat.server}
\alias{config.repeat.server}
\title{Repeat function handler for server_config}
\usage{
config.repeat.server(n = 1, ...)
}
\arguments{
\item{n}{Number of items}

\item{...}{Additional arguments}
}
\value{
list of server_config objects with the same parameters
}
\description{
Repeat function handler for server_config
}
