% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_request.r
\name{create_http_request}
\alias{create_http_request}
\title{Http request constructor}
\usage{
create_http_request()
}
\value{
http request trajectory
}
\description{
Create simmer trajectory wrapper for http request
}
