% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_config.r
\name{create_servers_configs}
\alias{create_servers_configs}
\title{Create multiple server configs}
\usage{
create_servers_configs(n = numeric(), ..., name_template = "Server-\%d")
}
\arguments{
\item{n}{Number of server configs to create}

\item{...}{Holds list of items}

\item{name_template}{Template for naming servers, must contain \%d to map number of server}
}
\value{
list of server configs
}
\description{
Create multiple server configs
}
