% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.extensions.r
\name{enrich_get_mon.policy_simulation_result}
\alias{enrich_get_mon.policy_simulation_result}
\title{S3 Method of 'enrich_get_mon' for 'policy_simulation_result' objects}
\usage{
\method{enrich_get_mon}{policy_simulation_result}(envs, get_mon_function)
}
\arguments{
\item{envs}{simmer environments}

\item{get_mon_function}{get_mon_* function from the family of monitoring functions}
}
\value{
Monitored data with enriched columns
}
\description{
S3 Method of 'enrich_get_mon' for 'policy_simulation_result' objects
}
