% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.extensions.r
\name{enrich_get_mon}
\alias{enrich_get_mon}
\title{Enrich method for get_mon_* results from simmer}
\usage{
enrich_get_mon(envs, get_mon_function)
}
\arguments{
\item{envs}{simmer environments}

\item{get_mon_function}{get_mon_* function from the family of monitoring functions}
}
\value{
Monitored data with enriched columns
}
\description{
Enrich method for get_mon_* results from simmer
}
