% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.extensions.r
\name{fix_cyrillic_rendering}
\alias{fix_cyrillic_rendering}
\title{Helper function to fix cyrillic symbols rendering on plots}
\usage{
fix_cyrillic_rendering()
}
\description{
Issues details: https://github.com/rstudio/rstudio/issues/2142
}
