% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{generate.scheduler_config}
\alias{generate.scheduler_config}
\title{Generate many scheduler configs with the same properties}
\usage{
generate.scheduler_config(n = 1, ...)
}
\arguments{
\item{n}{Number of copies}

\item{...}{Properties}
}
\value{
list of generated configs
}
\description{
Generate many scheduler configs with the same properties
}
