% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{generate.server_config}
\alias{generate.server_config}
\title{Generate many configs with the same properties}
\usage{
generate.server_config(n = 1, ...)
}
\arguments{
\item{n}{Number of copies}

\item{...}{Properties}
}
\value{
list of generated configs
}
\description{
Generate many configs with the same properties
}
