% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{generate_list_of_named_items.character}
\alias{generate_list_of_named_items.character}
\title{S3 method for character overload of generate_list_of_named_items function}
\usage{
\method{generate_list_of_named_items}{character}(x, name_template, create_function, ...)
}
\description{
S3 method for character overload of generate_list_of_named_items function
}
\keyword{internal}
