% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{generate_list_of_named_items.default}
\alias{generate_list_of_named_items.default}
\title{S3 method for default overload of generate_list_of_named_items function}
\usage{
\method{generate_list_of_named_items}{default}(x, name_template, create_function, ...)
}
\description{
S3 method for default overload of generate_list_of_named_items function
}
\keyword{.internal}
