% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{generate_list_of_named_items.numeric}
\alias{generate_list_of_named_items.numeric}
\title{Generate list of objects with name parameter based on creation function}
\usage{
\method{generate_list_of_named_items}{numeric}(x, name_template, create_function, ...)
}
\arguments{
\item{x}{input vector for names generation}

\item{name_template}{template for sprintf with \%s replace character}

\item{create_function}{creation function}

\item{...}{additional arguments}
}
\value{
list of object
}
\description{
All arguments with 'name' argument will be passed to creation function as 'prepared_args' list argument
}
