% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.utils.R
\name{generate_list_of_named_items}
\alias{generate_list_of_named_items}
\title{Generate list of named items according to passed creation function}
\usage{
generate_list_of_named_items(n = 1, name_template, create_function, ...)
}
\arguments{
\item{n}{Number of items to create}

\item{name_template}{Name template for sprintf}

\item{create_function}{Function which will be used for create}

\item{...}{Additional parameters required for specific function}
}
\description{
Generate list of named items according to passed creation function
}
\keyword{internal}
