% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_config.r
\name{generate_resource_config}
\alias{generate_resource_config}
\title{Generic method for creating typed configs with specific class name}
\usage{
generate_resource_config(..., class_name, prepared_args = list())
}
\arguments{
\item{...}{Holds list of items which simmer::add_resource requires. If any argument is unknown the error will be displayed.}

\item{class_name}{Class of resource config}

\item{prepared_args}{Primary args to use}
}
\value{
Typed resource config
}
\description{
Generic method for creating typed configs with specific class name
}
\keyword{internal}
