% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_distribution_function}
\alias{get_distribution_function}
\title{Convert vector of values from some column in data frame to distribution function
Original approach was taken from answer here:
https://stackoverflow.com/questions/32871602/r-generate-data-from-a-probability-density-distribution}
\usage{
get_distribution_function(data, column, only.value = T)
}
\arguments{
\item{data}{Provided data.frame}

\item{column}{Column name}

\item{only.value}{If is equal to TRUE then generator function will always return single value, otherwise it will return pairs of x and y}
}
\value{
function which accepts number and returns random value from generated distribution
}
\description{
Convert vector of values from some column in data frame to distribution function
Original approach was taken from answer here:
https://stackoverflow.com/questions/32871602/r-generate-data-from-a-probability-density-distribution
}
