% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mon_stats.r
\name{get_mon_arrival_summary}
\alias{get_mon_arrival_summary}
\title{Get summary from mon arrivals}
\usage{
get_mon_arrival_summary(.data, period = 1, units = "hours")
}
\arguments{
\item{.data}{arrivals data}

\item{period}{interarrival period}

\item{units}{interarrival period time unitss}
}
\description{
Get summary from mon arrivals
}
