% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulation_result.r
\name{get_mon_arrivals.policy_simulation_result}
\alias{get_mon_arrivals.policy_simulation_result}
\title{S3 Methods for get_mon functions}
\usage{
\method{get_mon_arrivals}{policy_simulation_result}(.envs, per_resource = FALSE, ongoing = FALSE)
}
\arguments{
\item{.envs}{vector of environments}

\item{per_resource}{Display information per resource}

\item{ongoing}{Display not finished tasks}
}
\value{
Enriched data
}
\description{
S3 Methods for get_mon functions
}
