% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.r
\name{get_policy_name.policy}
\alias{get_policy_name.policy}
\title{Get policy name}
\usage{
\method{get_policy_name}{policy}(policy)
}
\arguments{
\item{policy}{policy wrapper object}
}
\value{
policy name consumed by simmer
}
\description{
Get policy name
}
