% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.r
\name{get_policy_name}
\alias{get_policy_name}
\title{S3 generic for getting policy name}
\usage{
get_policy_name(policy)
}
\arguments{
\item{policy}{policy wrapper object}
}
\description{
S3 generic for getting policy name
}
