% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_policy.r
\name{get_resources_function.custom_policy}
\alias{get_resources_function.custom_policy}
\title{Get resource function for custom policy S3 method}
\usage{
\method{get_resources_function}{custom_policy}(policy)
}
\arguments{
\item{policy}{policy object}
}
\value{
resource function
}
\description{
In any "simmer" scenario there is no explicit way to define custom policy with custom any.
However, custom policy can be provided by function which produces a list of resources for selection.
If such function will provide only one value then any "simmer" policy function will only choose the one.
Technically, it means that custom function will provide "optimal" value for it's algorithm.
}
