% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.r
\name{get_resources_function.policy}
\alias{get_resources_function.policy}
\title{Get policy resources function}
\usage{
\method{get_resources_function}{policy}(policy)
}
\arguments{
\item{policy}{policy wrapper object}
}
\value{
policy function to be used
}
\description{
Get policy resources function
}
