% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.r
\name{get_resources_function}
\alias{get_resources_function}
\title{S3 generic for getting resources from policy}
\usage{
get_resources_function(policy)
}
\arguments{
\item{policy}{policy wrapper object}
}
\description{
S3 generic for getting resources from policy
}
