% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mon_stats.r
\name{get_windowed_mon_arrivals}
\alias{get_windowed_mon_arrivals}
\title{Extension for simmer which allows to get arrivals by time window of certain duration}
\usage{
get_windowed_mon_arrivals(
  .envs,
  period = 1,
  units = "hours",
  ongoing = FALSE,
  per_resource = FALSE
)
}
\arguments{
\item{.envs}{environments to get monitored data}

\item{period}{window time period}

\item{units}{window time unit}

\item{ongoing}{get ongoing tasks or no}

\item{per_resource}{display data per resource}
}
\description{
Extension for simmer which allows to get arrivals by time window of certain duration
}
