% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{handle_request_with_server}
\alias{handle_request_with_server}
\title{Handle method for request trajectory}
\usage{
handle_request_with_server(.trj, ...)
}
\arguments{
\item{.trj}{request trajectory}

\item{...}{
  Arguments passed on to \code{\link[=handle_request_with]{handle_request_with}}
  \describe{
    \item{\code{resource}}{resource name to seize and release}
    \item{\code{handlers}}{number of discrete resources to be acquired}
    \item{\code{handle_function}}{timeout function or handle time function}
  }}
}
\description{
Handle wrapper for server
}
