% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it_infrastructure.r
\name{is_it_infrastructure}
\alias{is_it_infrastructure}
\title{Validation of IT infrastructure}
\usage{
is_it_infrastructure(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE or FALSE
}
\description{
Validation of IT infrastructure
}
