% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulation_result.r
\name{is_policy_simulation_result}
\alias{is_policy_simulation_result}
\title{Check if object of class 'policy_simulation_result'}
\usage{
is_policy_simulation_result(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
TRUE or FALSE
}
\description{
Check if object of class 'policy_simulation_result'
}
