% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{is_request}
\alias{is_request}
\title{Request methods}
\usage{
is_request(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE or FALSE
}
\description{
Check whether object has class 'request' and is simmer trajectory
}
