% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_config.r
\name{is_resource_config}
\alias{is_resource_config}
\title{Check whether object is resource_config}
\usage{
is_resource_config(obj)
}
\arguments{
\item{obj}{Object to verify}
}
\value{
TRUE or FALSE
}
\description{
Check whether object is resource_config
}
