% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.extensions.r
\name{is_simmer}
\alias{is_simmer}
\title{Simmer extensions}
\usage{
is_simmer(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE or FALSE
}
\description{
Check if object is of class "simmer" or "wrap"
}
