% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_config.r
\name{is_simulation_config}
\alias{is_simulation_config}
\title{Simulation config type check}
\usage{
is_simulation_config(x)
}
\arguments{
\item{x}{Object to verify}
}
\value{
TRUE or FALSE
}
\description{
Simulation config type check
}
