% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.extensions.r
\name{is_trajectory}
\alias{is_trajectory}
\title{Simmer extensions}
\usage{
is_trajectory(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE or FALSE
}
\description{
Check whether object of class 'trajectory'
}
