% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it_infrastructure.r
\name{it_infrastructure}
\alias{it_infrastructure}
\title{Create IT infrastructure simulation environment}
\usage{
it_infrastructure(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[simmer:simmer]{simmer::simmer}}
  \describe{
    \item{\code{name}}{the name of the simulator.}
    \item{\code{verbose}}{enable showing activity information.}
    \item{\code{mon}}{monitor (in memory by default); see \code{\link[simmer]{monitor}} for
other options.}
    \item{\code{log_level}}{debugging level (see \code{\link[simmer]{log_}}).}
  }}
}
\value{
simmer environment with the name "IT-infrastructure"
}
\description{
Create IT infrastructure simulation environment
}
