% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{data}
\name{iti_metadata}
\alias{iti_metadata}
\title{Common metadata constants of IT infrastructure simulation framework}
\format{
An object of class \code{list} of length 3.
}
\usage{
iti_metadata
}
\description{
Common metadata constants of IT infrastructure simulation framework
}
\keyword{datasets}
