% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{data}
\name{iti_requests}
\alias{iti_requests}
\title{Common constants for IT infrastructure entities}
\format{
An object of class \code{list} of length 5.
}
\usage{
iti_requests
}
\description{
Common constants for IT infrastructure entities
}
\keyword{datasets}
