% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.extensions.r
\name{mask_resources}
\alias{mask_resources}
\title{Mask resources in simmer environment}
\usage{
mask_resources(.envs, values, mask)
}
\arguments{
\item{.envs}{simulation environments}

\item{values}{Vector of resource names to mask}

\item{mask}{Expected mask name}
}
\value{
Cloned environment with masked resources
}
\description{
Mask resources in simmer environment
}
