% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_policy.r
\name{new_custom_policy}
\alias{new_custom_policy}
\title{Custom policy}
\usage{
new_custom_policy(
  name = character(),
  resources = character(),
  env = NA,
  select_function = get_simmer_policy_resources,
  initial_container = list()
)
}
\arguments{
\item{name}{policy name}

\item{resources}{vector of resource names}

\item{env}{environment as a part of policy container, which can be used as a parameter in custom function}

\item{select_function}{resource selection function}

\item{initial_container}{container of values, from which policy object will be constructed}
}
\value{
new custom policy object
}
\description{
Custom policy
}
