% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{new_request}
\alias{new_request}
\title{Generic request constructor}
\usage{
new_request(name = character())
}
\arguments{
\item{name}{request name, like http, ftp, ...}
}
\value{
new simmer trajectory abstracted to IT request
}
\description{
Generic request constructor
}
