% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_config.r
\name{new_resource_config}
\alias{new_resource_config}
\title{Resource config object}
\usage{
new_resource_config(..., class = character())
}
\arguments{
\item{...}{Holds list of items}

\item{class}{Custom class name to apply to resulting object}
}
\value{
'resource_config' object with prepared named items
}
\description{
List of parameters for add_resource function from simmer package
}
