% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler_config.R
\name{new_scheduler_config}
\alias{new_scheduler_config}
\title{Create scheduler config}
\usage{
new_scheduler_config(..., prepared_args = list())
}
\arguments{
\item{...}{Holds list of items}

\item{prepared_args}{Primary args to use}
}
\value{
'scheduler config' object
}
\description{
Create scheduler config
}
