% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_config.r
\name{new_server_config}
\alias{new_server_config}
\title{Create server config}
\usage{
new_server_config(..., prepared_args = list())
}
\arguments{
\item{...}{Holds list of items}

\item{prepared_args}{Primary args to use}
}
\value{
'server config' object
}
\description{
Create server config
}
