% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cache.r
\name{new_simulation_cache}
\alias{new_simulation_cache}
\title{Simulation cache object}
\usage{
new_simulation_cache(env)
}
\arguments{
\item{env}{simmer environment}
}
\value{
new simulation cache object
}
\description{
Simulation cache object
}
