% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_config.r
\name{new_simulation_config}
\alias{new_simulation_config}
\title{Simulation config constructor}
\usage{
new_simulation_config(tasks = numeric(), until = double(), units = "second")
}
\arguments{
\item{tasks}{Number of tasks to produce}

\item{until}{Time until simulation should last}

\item{units}{Until time units}
}
\value{
New object of 'simulation_config'
}
\description{
Simulation config constructor
}
