% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulation_result.r
\name{pack_policy_simulation_result}
\alias{pack_policy_simulation_result}
\title{Pack policy simulation result into single environment if it is in list}
\usage{
pack_policy_simulation_result(x)
}
\arguments{
\item{x}{list with simulation environment and policy name}
}
\value{
environment with policy name
}
\description{
Pack policy simulation result into single environment if it is in list
}
