% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mon_stats.r
\name{parse_request_id}
\alias{parse_request_id}
\title{Extend mon arrivals or mon attributes with request ID}
\usage{
parse_request_id(.data, request_class)
}
\arguments{
\item{.data}{Monitor data obtained from simmer environment}

\item{request_class}{Request class whose ID's have to be parsed}
}
\description{
Extend mon arrivals or mon attributes with request ID
}
