% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.extensions.r
\name{plot_itin_statistics}
\alias{plot_itin_statistics}
\title{Plot IT infrastructure statistics}
\usage{
plot_itin_statistics(data, x = "time_period", y, group_var, legend_title)
}
\arguments{
\item{data}{data to use}

\item{x}{x axis variable name}

\item{y}{y axis variable name}

\item{group_var}{grouping variable name}

\item{legend_title}{legend title}
}
\value{
plot
}
\description{
Plot IT infrastructure statistics
}
