% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.r
\name{policy_initialize.policy}
\alias{policy_initialize.policy}
\title{Policy initialize function}
\usage{
\method{policy_initialize}{policy}(policy, ...)
}
\arguments{
\item{policy}{policy wrapper}

\item{...}{arguments to enrich with}
}
\value{
policy wrapper
}
\description{
Policy initialize function
}
