% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.r
\name{policy_initialize}
\alias{policy_initialize}
\title{S3 genric for policy initialization}
\usage{
policy_initialize(policy, ...)
}
\arguments{
\item{policy}{policy wrapper object}

\item{...}{additional parameters to enrich with}
}
\description{
S3 genric for policy initialization
}
