% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{prepare_args_from_caller}
\alias{prepare_args_from_caller}
\title{Prepare arguments from the list for function evaluation}
\usage{
prepare_args_from_caller(func, args)
}
\arguments{
\item{func}{Functions whose arguments have to be mapped}

\item{args}{Arguments list used for mapping}
}
\value{
Subset of arguments specific for function call
}
\description{
Prepare arguments from the list for function evaluation
}
