% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.r
\name{repeat_policy_scenario}
\alias{repeat_policy_scenario}
\title{Repeat same simulation with different resource-selection policies}
\usage{
repeat_policy_scenario(policies, scenario, sim_config)
}
\arguments{
\item{policies}{Policies to use in simulation}

\item{scenario}{Scenario to evaluate}

\item{sim_config}{Simulation config}
}
\value{
A list of \code{policy_simulation_results} type
}
\description{
Repeat same simulation with different resource-selection policies
}
