% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.r
\name{run_scenario}
\alias{run_scenario}
\title{Run IT infrastructure simulation scenario}
\usage{
run_scenario(scenario, policies, resource_configs, sim_config)
}
\arguments{
\item{scenario}{Scenario to run}

\item{policies}{Policies to use in simulation}

\item{resource_configs}{Resource configs which will be used for resource creation}

\item{sim_config}{Simulation config}
}
\value{
list of simulation results
}
\description{
Run IT infrastructure simulation scenario
}
