% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.extensions.r
\name{scale_discrete_manual_ext}
\alias{scale_discrete_manual_ext}
\title{Extension for ggplot2}
\usage{
scale_discrete_manual_ext(..., aesthetics = NULL, values = NULL, name)
}
\arguments{
\item{...}{key value pairs, where key is aesthetic name and value is a vector of values to use for this aesthetic}

\item{aesthetics}{vector of aesthetics names in case if dots are empty or if number of aesthetics must be limited}

\item{values}{named list of values for each aesthetic}

\item{name}{legend title}
}
\value{
list of discrete scales
}
\description{
scale_discrete_manual doesn't allow to use multiple aesthetics in one functions call.
This methods overcomes this limitation by specifying key value pair of aesthetic name and it's discrete values
}
