% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_request.r
\name{send_http_request}
\alias{send_http_request}
\title{Http request methods}
\usage{
send_http_request(.env, .trj, interarrival_time_generator)
}
\arguments{
\item{.env}{IT infrastructure environment}

\item{.trj}{request trajectory}

\item{interarrival_time_generator}{interarrival time generator}
}
\description{
Abstraction analogy for add_generator in simmer, which is applicable only to
IT infrastructure object and to request trajectory
}
