% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.r
\name{simmer_policy_name_match}
\alias{simmer_policy_name_match}
\title{Simmer default policy name match}
\usage{
simmer_policy_name_match(
  policyName = c("shortest-queue", "shortest-queue-available", "round-robin",
    "round-robin-available", "first-available", "random", "random-available")
)
}
\description{
Simmer default policy name match
}
\keyword{internal}
