% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_config.r
\name{simulation_config}
\alias{simulation_config}
\title{Simulation config}
\usage{
simulation_config(tasks = numeric(), until = double(), units = "seconds")
}
\arguments{
\item{tasks}{Number of tasks to produce}

\item{until}{Time until simulation should last}

\item{units}{Until time units}
}
\value{
Validated object of class 'simulation_config'
}
\description{
Simulation config
}
