% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.extensions.r
\name{theme_itinf}
\alias{theme_itinf}
\title{Custom theme for IT infrastructure statistics}
\usage{
theme_itinf(legend_title)
}
\arguments{
\item{legend_title}{legend title}
}
\value{
list of discrete scales and theme_bw
}
\description{
Custom theme for IT infrastructure statistics
}
