% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it_infrastructure.r
\name{validate_it_infrastructure}
\alias{validate_it_infrastructure}
\title{Validation of IT infrastructure}
\usage{
validate_it_infrastructure(x)
}
\arguments{
\item{x}{object to test}
}
\value{
same object as provided
}
\description{
Validation of IT infrastructure
}
