% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.r
\name{validate_iti_simulation_scenario}
\alias{validate_iti_simulation_scenario}
\title{Validate IT infrastructure simulation scenario function}
\usage{
validate_iti_simulation_scenario(x, ...)
}
\description{
Validate IT infrastructure simulation scenario function
}
\keyword{internal}
