% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.r
\name{validate_policy_name}
\alias{validate_policy_name}
\title{Validate text which is going to be used for naming policy}
\usage{
validate_policy_name(x)
}
\arguments{
\item{x}{input text}
}
\value{
input text if it is validated
}
\description{
Validate text which is going to be used for naming policy
}
\keyword{internal}
