% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.r
\name{validate_policy_scenario}
\alias{validate_policy_scenario}
\title{Validate policy scenario function}
\usage{
validate_policy_scenario(x)
}
\description{
Validate policy scenario function
}
\keyword{internal}
