% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulation_result.r
\name{validate_policy_simulation_result}
\alias{validate_policy_simulation_result}
\title{Validate policy_simulation_result object}
\usage{
validate_policy_simulation_result(x)
}
\arguments{
\item{x}{Object to validate}
}
\description{
Validate policy_simulation_result object
}
