% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{validate_request}
\alias{validate_request}
\title{Request methods}
\usage{
validate_request(x)
}
\arguments{
\item{x}{object to validate}
}
\description{
Validate that object is a request and fail if it is not
}
