% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_config.r
\name{validate_resource_config}
\alias{validate_resource_config}
\title{Validate resource config object}
\usage{
validate_resource_config(obj)
}
\arguments{
\item{obj}{object to validate}
}
\value{
same object
}
\description{
Validate resource config object
}
\details{
Checks if object is of type resource_config and if not will stop execution
}
