% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.extensions.r
\name{validate_simmer}
\alias{validate_simmer}
\title{Simmer extensions}
\usage{
validate_simmer(x)
}
\arguments{
\item{x}{object to test}
}
\value{
same object as was provided
}
\description{
Validate object as a simmer object, otherwise fail
}
