% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.extensions.r
\name{validate_simmer_with_name}
\alias{validate_simmer_with_name}
\title{Simmer extensions}
\usage{
validate_simmer_with_name(x, assertedName)
}
\arguments{
\item{x}{Object to test}

\item{assertedName}{Expected name of simmer object}
}
\value{
TRUE or FALSE
}
\description{
Validate if object is of class "simmer" and it's name is equal to asertee name
}
