% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cache.r
\name{validate_simulation_cache}
\alias{validate_simulation_cache}
\title{Validate simulation cache object}
\usage{
validate_simulation_cache(x)
}
\arguments{
\item{x}{object to test}
}
\value{
invisible object or validation message
}
\description{
Validate simulation cache object
}
