% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_config.r
\name{validate_simulation_config}
\alias{validate_simulation_config}
\title{Validate simulation config object}
\usage{
validate_simulation_config(x)
}
\arguments{
\item{x}{Object to validate}
}
\description{
Validate simulation config object
}
