% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_config.r
\name{validate_typed_list_of_resource_configs}
\alias{validate_typed_list_of_resource_configs}
\title{Validate input object is typed list with type 'resource_config'}
\usage{
validate_typed_list_of_resource_configs(obj)
}
\arguments{
\item{obj}{object to validate}
}
\value{
validated object
}
\description{
Validate input object is typed list with type 'resource_config'
}
