% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_policy.r
\name{wrap_policy}
\alias{wrap_policy}
\title{Wrap string or function to a policy object}
\usage{
wrap_policy(...)
}
\arguments{
\item{...}{either default simmer policy name, like "shortest-queue", "random", etc. or custom function which returns resources to select}
}
\value{
a list of policy objects
}
\description{
Wrap string or function to a policy object
}
