% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{as.data.frame.density}
\alias{as.data.frame.density}
\title{S3 method for as.data.frame when 'density' type is provided}
\usage{
\method{as.data.frame}{density}(x, row.names, optional, ...)
}
\arguments{
\item{x}{any \R object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
data frame which contains only 'x' and 'y' values from 'density' object
}
\description{
S3 method for as.data.frame when 'density' type is provided
}
