% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it_infrastructure.r
\name{it_infrastructure}
\alias{it_infrastructure}
\title{Create IT infrastructure simulation environment}
\usage{
it_infrastructure(...)
}
\value{
simmer environment with the name "IT-infrastructure"
}
\description{
Create IT infrastructure simulation environment
}
